Undo non-static inner interface hackery.
Jeroen told me how to fix IKVM, which I did and I regenerated the stubs with
properly marked inner interfaces.

For anyone who got to this party early, you're going to have to

  rm -rf ~/.m2/repository/com/xamarin/monotouch

because I didn't feel like bumping the version numbers on the regenerated stub
jars. I promise not to be so lazy in the future.